requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
SQLAlchemy==2.0.0
python-dotenv==0.19.0
gunicorn==20.1.0
nltk==3.6.2
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
psycopg2-binary==2.9.5
python-magic==0.4.27
Pillow==8.3.1 